requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.7
djangorestframework==3.14.0
Faker==21.0.0
mysqlclient==2.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
